device/bridge/requirements.txt
pyserial==3.5
python-dotenv==1.1.0
requests>=2.31.0
supabase==2.15.2
